[ember] @ember/controller as source of types, ember as re-export
