+++ types/ember-data/tsconfig.json
@@ -18,6 +18,7 @@
             "@ember/debug/*": ["ember__debug/*"],
             "@ember/array": ["ember__array"],
             "@ember/array/*": ["ember__array/*"],
+            "@ember/controller": ["ember__controller"],
             "@ember/object": ["ember__object"],
             "@ember/object/*": ["ember__object/*"]
         },
