+++ types/ember__object/tsconfig.json
@@ -36,6 +36,7 @@
         "proxy.d.ts",
         "-private/types.d.ts",
         "-private/copyable.d.ts",
+        "-private/action-handler.d.ts",
         "test/lib/assert.ts",
         "test/access-modifier.ts",
         "test/core.ts",
